requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file> -c defaults -c conda-forge -c anaconda -c pytorch
# platform: win-64
# After this, you need to install torch_stft from pip:
# $ pip install torch_stft

# For AudioVisualizer
matplotlib=3.3.2
numpy=1.19.2

# For microphone and speakers
python-sounddevice=0.4.1

# For run script
pyyaml=5.4.1

# For DCCRN processor
librosa=0.6.3

pytorch=1.7.1
torchvision=0.8.2
torchaudio=0.7.2

cudatoolkit=10.1 # Use a version that works with the pytorch version and is installed on the computer
numba=0.48
ffmpeg

# For file reader and writer
pysoundfile=0.10.3